third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-shorthand-serialization-expected.txt

This is a testharness.js-based test.
[FAIL] multiple backgrounds with varying values
  assert_equals: expected "url(\\"/favicon.ico\\") left top no-repeat, url(\\"/favicon.ico\\") center center / 100% 100% no-repeat, white url(\\"/favicon.ico\\")" but got "url(\\"/favicon.ico\\") left top no-repeat, url(\\"/favicon.ico\\") center center / 100% 100% no-repeat, url(\\"/favicon.ico\\") white"
[FAIL] all initial values
  assert_equals: expected "none" but got "padding-box border-box"
Harness: the test ran to completion.
